Add files via upload
The below graph illustrates Las Vegas having a lot of establishments with ratings of 3 stars or more. This analysis led me to recommend Las Vegas as the destination to travel to. 


 

Furthermore, I ran the data into a k-nearest neighbors/means code, and based on the test and training data, we can predict Vegas, is the destination to pick based on the number of venues and number of venues with 3 stars or more. 

Train set: (87283, 3) (87283,)
Test set: (87284, 3) (87284,)
Train set Accuracy:  0.9831009474926389
Test set Accuracy:  0.9707162824801796